refactor: move types to dedicated folder
